Plugin Directory: Gate 2FA-related calls behind class_exists/function_exists checks to prevent fatal errors when Two_Factor plugin is not active.
  - Skip 2FA revalidation in release confirmation API when Two_Factor_Core class is unavailable
  - Wrap Two_Factor_Core::is_user_using_two_factor() calls in release confirmation shortcode
  - Guard user_requires_2fa() and is_email_address_unsafe() calls in upload shortcode

This allows for the plugin directory to be used in a local environment.

See https://github.com/WordPress/wordpress.org/pull/555


git-svn-id: https://meta.svn.wordpress.org/sites/trunk@14698 74240141-8908-4e6f-9713-ba540dce6ec7

+++ wordpress.org/public_html/wp-content/plugins/plugin-directory/shortcodes/class-upload.php
@@ -64,6 +64,8 @@ class Upload {
 
 		// Require 2FA for plugin authors on upload.
 		if (
+			function_exists( 'WordPressdotorg\Two_Factor\user_requires_2fa' ) &&
+			class_exists( 'Two_Factor_Core' ) &&
 			user_requires_2fa( wp_get_current_user() ) &&
 			! Two_Factor_Core::is_user_using_two_factor( get_current_user_id() )
 		) {
@@ -393,7 +395,7 @@ class Upload {
 					'https://wordpress.org/news/2024/12/holiday-break/'
 				)
 			);
-		} else if ( is_email_address_unsafe( wp_get_current_user()->user_email ) ) {
+		} else if ( function_exists( 'is_email_address_unsafe' ) /* multisite-only */ && is_email_address_unsafe( wp_get_current_user()->user_email ) ) {
 			echo '<div class="notice notice-error notice-alt"><p>' .
 				sprintf(
 					/* translators: %s: Profile edit url. */
